hbs/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
django-daraja==1.3.0
django-jazzmin==3.0.0
django-nested-admin==4.1.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==23.0.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
joblib==1.4.2
ndg-httpsclient==0.5.1
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
phonenumbers==8.13.43
pillow==10.4.0
priority==1.3.0
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-monkey-business==1.1.0
pytz==2024.2
reportlab==4.2.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
stripe==10.7.0
threadpoolctl==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
zope.interface==7.0.1
